PID Works well, More robust bodyconfig
I implemented a PID on either axis instead of combining them on one. This method works far better. There is a warper class that allows them to be modified a the same time. Additionally, I fixed the head height being off and the player movement at low croutch. The headset height was changes in RigXRTransformer and the "hip drag" in body config