Down-look spazzing fixed
A process like in bonelab where the player looks down and their torso remains in the same orientation of their heads is not implemented, but this works well enough. Though, it wouldnt be too hard to achi8eve as I already obtained a similar effect by simply not adjusting the y-axis past a certain x-axis rotation. I could create 3 "zones." One of free movement(normal), one of no movement(very low), and one of adjustment(sort of low where lerping happens) between the two. The current solution is comparatively simple. Ill implement the more bonelab style fix if it becomes apparent during testing that the more complex solution is preferable.