grab-priority for ownership
The transitions are still jittery, but--because the grabber of the object is preferred--the object no long "oscillates" between owners. The physics syncing is much cleaner, but I think it could do with an ownership transitioner that does something like--instead of taking the rigidbody data at time of transition in the new owner's perspective--wait a bit to account for delay in the simulations+some sort of smooth transition.